fix 64 bit mul long double algo the 3rd time

+++ 配套光盘/例题/0x00 基本算法/0x01 位运算/64位整数乘法/CH0102 64位整数乘法 方法二.cpp
@@ -10,7 +10,7 @@ ull mul(ull a, ull b, ull p) {
 	a %= p, b %= p;  // 当a,b一定在0~p之间时，此行不必要
 	ull c = (long double)a * b / p;
 	ull x = a * b, y = c * p;
-	long long ans = x % p - y % p;
+	long long ans = (long long)(x % p) - (long long)(y % p);
 	if (ans < 0) ans += p;
 	return ans;
 }
